refactor(core): consolidate types and formatting into plainDate module
- Move PlainDate interface to dateTypes.ts alongside ISODateString
- Move formatDisplayDate into plainDateFormatDisplay in plainDate.ts
- Use plainDateToDate in plainDateGetWeekNumber (remove unnecessary UTC)
- Consolidate dateParser createPlainDate to use plainDateCreate via try/catch
